requirements.txt
PyJWT~=1.5.0
Pillow~=8.3.1